# Changelog optimization