--- a/pkg/changelog/changelog.go
+++ b/pkg/changelog/changelog.go
@@ -1,7 +1,6 @@
 package changelog
 
 import (
-	"slices"
 	"time"
 
 	"github.com/aquasecurity/tracee/pkg/logger"
@@ -165,33 +164,57 @@ func (clv *Changelog[T]) findIndex(target time.Time) int {
 
 // enforceSizeBoundary ensures that the size of the inner array doesn't exceed the limit.
 // It applies two methods to reduce the log size to the maximum allowed:
-// 1. Unite duplicate values that are trailing one another.
+// 1. Unite duplicate values that are trailing one another, removing the oldest of the pair.
 // 2. Remove the oldest logs as they are likely less important.
 
 func (clv *Changelog[T]) enforceSizeBoundary() {
-	if len(clv.changes) > clv.maxSize {
-		// Get rid of oldest changes to keep max size boundary
-		boundaryDiff := len(clv.changes) - clv.maxSize
-
-		// First try to unite states
-		clv.changes = slices.CompactFunc(clv.changes, func(i item[T], i2 item[T]) bool {
-			if i.value == i2.value && boundaryDiff > 0 {
-				delete(clv.timestamps, i2.timestamp)
-				boundaryDiff--
-				return true
-			}
-			return false
-		})
-
-		if boundaryDiff == 0 {
-			return
-		}
-		removedChanges := clv.changes[:boundaryDiff]
-		clv.changes = clv.changes[boundaryDiff:]
-		for _, removedChange := range removedChanges {
-			delete(clv.timestamps, removedChange.timestamp)
-		}
+	if len(clv.changes) <= clv.maxSize {
+		// Nothing to do
+		return
 	}
+
+	boundaryDiff := len(clv.changes) - clv.maxSize
+	changed := false
+
+	// Compact the slice in place
+	writeIdx := 0
+	for readIdx := 0; readIdx < len(clv.changes); readIdx++ {
+		nextIdx := readIdx + 1
+		if nextIdx < len(clv.changes) &&
+			clv.changes[nextIdx].value == clv.changes[readIdx].value &&
+			boundaryDiff > 0 {
+			// Remove the oldest (readIdx) from the duplicate pair
+			delete(clv.timestamps, clv.changes[readIdx].timestamp)
+			boundaryDiff--
+			changed = true
+			continue
+		}
+
+		// If elements have been removed or moved, update the map and the slice
+		if changed {
+			clv.changes[writeIdx] = clv.changes[readIdx]
+		}
+
+		writeIdx++
+	}
+
+	if changed {
+		clear(clv.changes[writeIdx:])
+		clv.changes = clv.changes[:writeIdx]
+	}
+
+	if len(clv.changes) <= clv.maxSize {
+		// Size is within limits after compaction
+		return
+	}
+
+	// As it still exceeds maxSize, remove the oldest entries in the remaining slice
+	boundaryDiff = len(clv.changes) - clv.maxSize
+	for i := 0; i < boundaryDiff; i++ {
+		delete(clv.timestamps, clv.changes[i].timestamp)
+	}
+	clear(clv.changes[:boundaryDiff])
+	clv.changes = clv.changes[boundaryDiff:]
 }
 
 // returnZero returns the zero value of the type T.