--- a/pkg/changelog/changelog_test.go
+++ b/pkg/changelog/changelog_test.go
@@ -1,26 +1,25 @@
-package changelog_test
+package changelog
 
 import (
+	"reflect"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
-
-	"github.com/aquasecurity/tracee/pkg/changelog"
 )
 
 func TestChangelog(t *testing.T) {
 	t.Parallel()
 
 	t.Run("GetCurrent on an empty changelog", func(t *testing.T) {
-		cl := changelog.NewChangelog[int](3)
+		cl := NewChangelog[int](3)
 
 		// Test GetCurrent on an empty changelog
 		assert.Zero(t, cl.GetCurrent())
 	})
 
 	t.Run("Set and get", func(t *testing.T) {
-		cl := changelog.NewChangelog[int](3)
+		cl := NewChangelog[int](3)
 		testVal := 42
 
 		cl.SetCurrent(testVal)
@@ -28,7 +27,7 @@ func TestChangelog(t *testing.T) {
 	})
 
 	t.Run("Set and get on set time", func(t *testing.T) {
-		cl := changelog.NewChangelog[int](3)
+		cl := NewChangelog[int](3)
 		testVal1 := 42
 		testVal2 := 76
 		testVal3 := 76
@@ -50,7 +49,7 @@ func TestChangelog(t *testing.T) {
 	})
 
 	t.Run("Set twice on the same time", func(t *testing.T) {
-		cl := changelog.NewChangelog[int](3)
+		cl := NewChangelog[int](3)
 		testVal := 42
 
 		now := time.Now()
@@ -62,13 +61,13 @@ func TestChangelog(t *testing.T) {
 	})
 
 	t.Run("Get on an empty changelog", func(t *testing.T) {
-		cl := changelog.NewChangelog[int](3)
+		cl := NewChangelog[int](3)
 
 		assert.Zero(t, cl.GetCurrent())
 	})
 
 	t.Run("Test 1 second interval among changes", func(t *testing.T) {
-		cl := changelog.NewChangelog[int](3)
+		cl := NewChangelog[int](3)
 
 		cl.SetCurrent(1)
 		time.Sleep(2 * time.Second)
@@ -84,7 +83,7 @@ func TestChangelog(t *testing.T) {
 	})
 
 	t.Run("Test 100 milliseconds interval among changes", func(t *testing.T) {
-		cl := changelog.NewChangelog[int](3)
+		cl := NewChangelog[int](3)
 
 		cl.SetCurrent(1)
 		time.Sleep(100 * time.Millisecond)
@@ -100,7 +99,7 @@ func TestChangelog(t *testing.T) {
 	})
 
 	t.Run("Test getting all values at once", func(t *testing.T) {
-		cl := changelog.NewChangelog[int](3)
+		cl := NewChangelog[int](3)
 
 		cl.SetCurrent(1)
 		time.Sleep(100 * time.Millisecond)
@@ -113,7 +112,7 @@ func TestChangelog(t *testing.T) {
 	})
 
 	t.Run("Pass max size wit repeated values", func(t *testing.T) {
-		cl := changelog.NewChangelog[int](3)
+		cl := NewChangelog[int](3)
 
 		cl.SetCurrent(1)
 		time.Sleep(100 * time.Millisecond)
@@ -125,14 +124,14 @@ func TestChangelog(t *testing.T) {
 
 		now := time.Now()
 		assert.Equal(t, 1, cl.Get(now.Add(-300*time.Millisecond)))
-		assert.Equal(t, 2, cl.Get(now.Add(-200*time.Millisecond)))
+		assert.Equal(t, 1, cl.Get(now.Add(-200*time.Millisecond))) // oldest 2 is removed, so 1 is returned
 		assert.Equal(t, 2, cl.Get(now.Add(-100*time.Millisecond)))
 		assert.Equal(t, 3, cl.Get(now))
 		assert.Len(t, cl.GetAll(), 3)
 	})
 
 	t.Run("Pass max size with unique values", func(t *testing.T) {
-		cl := changelog.NewChangelog[int](3)
+		cl := NewChangelog[int](3)
 
 		cl.SetCurrent(1)
 		time.Sleep(100 * time.Millisecond)
@@ -151,7 +150,7 @@ func TestChangelog(t *testing.T) {
 	})
 
 	t.Run("Pass max size with new old value", func(t *testing.T) {
-		cl := changelog.NewChangelog[int](3)
+		cl := NewChangelog[int](3)
 
 		cl.SetCurrent(1)
 		time.Sleep(100 * time.Millisecond)
@@ -173,7 +172,7 @@ func TestChangelog(t *testing.T) {
 	})
 
 	t.Run("Zero sized changelog", func(t *testing.T) {
-		cl := changelog.NewChangelog[int](0)
+		cl := NewChangelog[int](0)
 
 		cl.SetCurrent(1)
 		time.Sleep(100 * time.Millisecond)
@@ -193,4 +192,159 @@ func TestChangelog(t *testing.T) {
 		assert.Equal(t, 0, cl.Get(now))
 		assert.Empty(t, cl.GetAll())
 	})
+
+	t.Run("Test enforceSizeBoundary", func(t *testing.T) {
+		type TestCase struct {
+			name               string
+			maxSize            int
+			initialChanges     []item[int]
+			expectedChanges    []item[int]
+			expectedTimestamps map[time.Time]struct{}
+		}
+
+		testCases := []TestCase{
+			{
+				name:    "No Action Required",
+				maxSize: 3,
+				initialChanges: []item[int]{
+					{timestamp: getTimeFromSec(42), value: 1},
+					{timestamp: getTimeFromSec(43), value: 2},
+					{timestamp: getTimeFromSec(44), value: 3},
+				},
+				expectedChanges: []item[int]{
+					{timestamp: getTimeFromSec(42), value: 1},
+					{timestamp: getTimeFromSec(43), value: 2},
+					{timestamp: getTimeFromSec(44), value: 3},
+				},
+				expectedTimestamps: map[time.Time]struct{}{
+					getTimeFromSec(42): {},
+					getTimeFromSec(43): {},
+					getTimeFromSec(44): {},
+				},
+			},
+			{
+				name:    "Basic Removal of Oldest Entries",
+				maxSize: 3,
+				initialChanges: []item[int]{
+					{timestamp: getTimeFromSec(42), value: 1},
+					{timestamp: getTimeFromSec(43), value: 2},
+					{timestamp: getTimeFromSec(44), value: 3},
+					{timestamp: getTimeFromSec(45), value: 4},
+				},
+				expectedChanges: []item[int]{
+					{timestamp: getTimeFromSec(43), value: 2},
+					{timestamp: getTimeFromSec(44), value: 3},
+					{timestamp: getTimeFromSec(45), value: 4},
+				},
+				expectedTimestamps: map[time.Time]struct{}{
+					getTimeFromSec(43): {},
+					getTimeFromSec(44): {},
+					getTimeFromSec(45): {},
+				},
+			},
+			{
+				name:    "Compacting Duplicate Values - Start",
+				maxSize: 3,
+				initialChanges: []item[int]{
+					{timestamp: getTimeFromSec(42), value: 1},
+					{timestamp: getTimeFromSec(43), value: 1},
+					{timestamp: getTimeFromSec(44), value: 2},
+					{timestamp: getTimeFromSec(45), value: 3},
+				},
+				expectedChanges: []item[int]{
+					{timestamp: getTimeFromSec(43), value: 1},
+					{timestamp: getTimeFromSec(44), value: 2},
+					{timestamp: getTimeFromSec(45), value: 3},
+				},
+				expectedTimestamps: map[time.Time]struct{}{
+					getTimeFromSec(43): {},
+					getTimeFromSec(44): {},
+					getTimeFromSec(45): {},
+				},
+			},
+			{
+				name:    "Compacting Duplicate Values - Middle",
+				maxSize: 3,
+				initialChanges: []item[int]{
+					{timestamp: getTimeFromSec(42), value: 1},
+					{timestamp: getTimeFromSec(43), value: 2},
+					{timestamp: getTimeFromSec(44), value: 2},
+					{timestamp: getTimeFromSec(45), value: 3},
+				},
+				expectedChanges: []item[int]{
+					{timestamp: getTimeFromSec(42), value: 1},
+					{timestamp: getTimeFromSec(44), value: 2},
+					{timestamp: getTimeFromSec(45), value: 3},
+				},
+				expectedTimestamps: map[time.Time]struct{}{
+					getTimeFromSec(42): {},
+					getTimeFromSec(44): {},
+					getTimeFromSec(45): {},
+				},
+			},
+			{
+				name:    "Compacting Duplicate Values - End",
+				maxSize: 3,
+				initialChanges: []item[int]{
+					{timestamp: getTimeFromSec(42), value: 1},
+					{timestamp: getTimeFromSec(43), value: 2},
+					{timestamp: getTimeFromSec(44), value: 3},
+					{timestamp: getTimeFromSec(45), value: 3},
+				},
+				expectedChanges: []item[int]{
+					{timestamp: getTimeFromSec(42), value: 1},
+					{timestamp: getTimeFromSec(43), value: 2},
+					{timestamp: getTimeFromSec(45), value: 3},
+				},
+				expectedTimestamps: map[time.Time]struct{}{
+					getTimeFromSec(42): {},
+					getTimeFromSec(43): {},
+					getTimeFromSec(45): {},
+				},
+			},
+			{
+				name:    "Combination of Compaction and Removal of Oldest Entries",
+				maxSize: 3,
+				initialChanges: []item[int]{
+					{timestamp: getTimeFromSec(42), value: 1},
+					{timestamp: getTimeFromSec(43), value: 2},
+					{timestamp: getTimeFromSec(44), value: 2},
+					{timestamp: getTimeFromSec(45), value: 2},
+					{timestamp: getTimeFromSec(46), value: 3},
+					{timestamp: getTimeFromSec(47), value: 4},
+				},
+				expectedChanges: []item[int]{
+					{timestamp: getTimeFromSec(45), value: 2},
+					{timestamp: getTimeFromSec(46), value: 3},
+					{timestamp: getTimeFromSec(47), value: 4},
+				},
+				expectedTimestamps: map[time.Time]struct{}{
+					getTimeFromSec(45): {},
+					getTimeFromSec(46): {},
+					getTimeFromSec(47): {},
+				},
+			},
+		}
+
+		for _, tc := range testCases {
+			t.Run(tc.name, func(t *testing.T) {
+				cl := NewChangelog[int](tc.maxSize)
+				for _, change := range tc.initialChanges {
+					cl.Set(change.value, change.timestamp)
+				}
+
+				cl.enforceSizeBoundary()
+
+				eq := reflect.DeepEqual(cl.timestamps, tc.expectedTimestamps)
+				assert.True(t, eq)
+
+				eq = reflect.DeepEqual(cl.changes, tc.expectedChanges)
+				assert.True(t, eq)
+			})
+		}
+	})
+}
+
+func getTimeFromSec(second int) time.Time {
+	return time.Unix(int64(second), 0)
 }