### 1. Explain what the PR does

8cee7a26f **perf(changelog): optimize enforceSizeBoundary**
3bc61a882 **chore(changelog): avoid unnecessary initialization**


8cee7a26f **perf(changelog): optimize enforceSizeBoundary**

```
These optimizations make the enforceSizeBoundary function more
efficient across various scenarios, leading to overall better
performance while maintaining the same behaviour (*).

Running tool: /home/gg/.goenv/versions/1.22.4/bin/go test -benchmem
-run=^$ -tags ebpf
-bench ^(BenchmarkEnforceSizeBoundaryOld|BenchmarkEnforceSizeBoundary)$
github.com/aquasecurity/tracee/pkg/changelog -benchtime=100000000x

goos: linux
goarch: amd64
pkg: github.com/aquasecurity/tracee/pkg/changelog
cpu: AMD Ryzen 9 7950X 16-Core Processor

| Test Case              | Old (ns/op) | New (ns/op) |   (%)  |
|------------------------|-------------|-------------|--------|
| No change needed       | 1.446       | 1.434       |  0.83% |
| Trim excess duplicates | 30.18       | 21.97       | 27.21% |
| Remove oldest entries  | 27.29       | 21.39       | 21.60% |

Since enforceSizeBoundary is called by setAt, the performance
improvements will be reflected in the latter:

-bench ^(setAtOld|setAt)$

| Test Case          | Old (ns/op) | New (ns/op) |   (%)  |
|--------------------|-------------|-------------|--------|
| All Scenarios      | 1267        | 1163        | 8.21%  |
| Within Limit       | 1770        | 1747        | 1.30%  |

Tests were fixed and added to ensure the correctness of the
implementation. Benchmarks were also added to track the performance
improvements.

(*) The behaviour of the function is the same, but the implementation
fixes a bug that when coalescing duplicate values, the removal of the
oldest entry was not being done correctly. Instead of removing the
oldest entry, the function was removing the newest one.
```

### 2. Explain how to test it

<!--
Maintainer will review the code, and test the fix/feature, how to run Tracee ?
Give a full command line example and what to look for.
-->

### 3. Other comments

This is part of the proctree effort.