@geyslan 2024-08-10
Converted it as draft since the `setAt` optimization is demanding changes on `enforceSizeBoundary` as well. So the benchmark results are jeopardised.