@geyslan 2024-08-23
Probably related: https://github.com/aquasecurity/tracee/issues/3878