@geyslan 2024-08-23
> LGTM, @geyslan just double check the setAt because couldn't run it here.

Finished (10000000x):

```
Running tool: /home/gg/.goenv/versions/1.22.4/bin/go test -benchmem -run=^$ -tags ebpf -bench ^Benchmark_setAt$ github.com/aquasecurity/tracee/pkg/changelog -benchtime=10000000x

goos: linux
goarch: amd64
pkg: github.com/aquasecurity/tracee/pkg/changelog
cpu: AMD Ryzen 9 7950X 16-Core Processor            
Benchmark_setAt/All_Scenarios-32         	10000000	      1187 ns/op	    1008 B/op	       7 allocs/op
Benchmark_setAt/Within_Limit-32          	10000000	      1586 ns/op	    2571 B/op	       8 allocs/op
```